ROVR_data_decryptor/requirements.txt
pycryptodome>=3.15.0
tqdm>=4.64.0
